Aggregate totals by total method time per endpoint call
Calculates the mean time for methods based on the number of calls to the
endpoint from which the method call was made as opposed to just
calculating the mean method call time without regard for the number of
surrounding endpoint calls from which the method calls were made.

Also reworks the output for the base vs. compare files to list
information correctly for the base aggregation results.  Previously, the
"compare" results were being listed instead of the "base" results for
the "base" output files.